Add files via upload
This program will replace all occurrences of a word "ABC" provided by the user    
with the word "DEF", also provided by the user and only those lines where the condition word "MNO" is located (also word provided by the user).                                                          
For this example, we will use a specific file called  "text_example.txt" provided in this repository as well.